Add search_papers_from_airas_db
