+++ src/airas/features/retrieve/retrieve_conference_paper_from_query_subgraph/retrieve_conference_paper_from_query_subgraph.py
@@ -13,14 +13,14 @@ from airas.core.base import BaseSubgraph
 from airas.features.retrieve.nodes.extract_github_url_from_text import (
     extract_github_url_from_text,
 )
-from airas.features.retrieve.nodes.get_filtered_papers_from_conference import (
-    get_filterd_papers_from_conference,
-)
 from airas.features.retrieve.nodes.retrieve_arxiv_text_from_url import (
     retrieve_arxiv_text_from_url,
 )
 from airas.features.retrieve.nodes.search_arxiv import (
     search_arxiv,
+)
+from airas.features.retrieve.nodes.search_papers_from_airas_db import (
+    search_papers_from_airas_db,
 )
 from airas.features.retrieve.nodes.select_best_paper import (
     select_best_paper,
@@ -90,14 +90,12 @@ class RetrieveConferencePaperFromQuerySubgraph(BaseSubgraph):
         self,
         llm_name: str,
         save_dir: str,
-        paper_json_urls: list[str],
         arxiv_query_batch_size: int = 10,
         arxiv_num_retrieve_paper: int = 1,
         arxiv_period_days: int | None = None,
     ):
         self.llm_name = llm_name
         self.save_dir = save_dir
-        self.paper_json_urls = paper_json_urls
         self.arxiv_query_batch_size = arxiv_query_batch_size
         self.arxiv_num_retrieve_paper = arxiv_num_retrieve_paper
         self.arxiv_period_days = arxiv_period_days
@@ -121,11 +119,10 @@ class RetrieveConferencePaperFromQuerySubgraph(BaseSubgraph):
         }
 
     @retrieve_paper_from_query_timed
-    def _get_filtered_papers_from_conference_node(
+    def _search_papers_from_airas_db(
         self, state: RetrieveConferencePaperFromQueryState
     ) -> dict[str, list[str]]:
-        extracted_paper_titles = get_filterd_papers_from_conference(
-            json_urls=self.paper_json_urls,
+        extracted_paper_titles = search_papers_from_airas_db(
             queries=state["base_queries"],
         )
         return {"extracted_paper_titles": extracted_paper_titles}
@@ -294,8 +291,8 @@ class RetrieveConferencePaperFromQuerySubgraph(BaseSubgraph):
 
         graph_builder.add_node("initialize_state", self._initialize_state)
         graph_builder.add_node(
-            "get_filtered_papers_from_conference",
-            self._get_filtered_papers_from_conference_node,
+            "search_papers_from_airas_db",
+            self._search_papers_from_airas_db,
         )
         graph_builder.add_node(
             "search_arxiv_node", self._search_arxiv_node
@@ -311,11 +308,9 @@ class RetrieveConferencePaperFromQuerySubgraph(BaseSubgraph):
         graph_builder.add_node("prepare_state", self._prepare_state)
 
         graph_builder.add_edge(START, "initialize_state")
-        graph_builder.add_edge(
-            "initialize_state", "get_filtered_papers_from_conference"
-        )
+        graph_builder.add_edge("initialize_state", "search_papers_from_airas_db")
         graph_builder.add_conditional_edges(
-            source="get_filtered_papers_from_conference",
+            source="search_papers_from_airas_db",
             path=self._check_extracted_titles,
             path_map={
                 "Stop": END,
@@ -349,13 +344,6 @@ class RetrieveConferencePaperFromQuerySubgraph(BaseSubgraph):
 
 
 def main():
-    JSON_URLS = [
-        "https://icml.cc/static/virtual/data/icml-2024-orals-posters.json",
-        "https://iclr.cc/static/virtual/data/iclr-2024-orals-posters.json",
-        "https://nips.cc/static/virtual/data/neurips-2024-orals-posters.json",
-        "https://cvpr.thecvf.com/static/virtual/data/cvpr-2024-orals-posters.json",
-    ]
-
     llm_name = "o3-mini-2025-01-31"
     save_dir = "/workspaces/airas/data"
     input = retrieve_paper_from_query_subgraph_input_data
@@ -366,7 +354,6 @@ def main():
     result = RetrieveConferencePaperFromQuerySubgraph(
         llm_name=llm_name,
         save_dir=save_dir,
-        paper_json_urls=JSON_URLS,
     ).run(input)
     print(f"result: {json.dumps(result, indent=2)}")
